+++ CHANGELOG.md
@@ -1,5 +1,6 @@
 0.8.2 - dev
 - [ ] literal rank of writer depending of num of com
+- [ ] add dashboard module
 
 0.8.1 - 2021.09.17
 - fix plurals and better translation